Translation of all descriptions to buildings and Room Owerlay
Translated all other descriptions of buildings on the tabs: stations, utilities, auto, delivery.
Changes in rectangles are some elements of construction.
Fully translated Overliy Rooms and the entire description of the rooms and what is needed for their full functioning.
Translated some small lines of the code into the translation file, but a small number.