Translated, update readme, backup
Translated all the random names of the world when creating the world. Now the names of the duplicates are also translated, will be tested in the near future. All messages that appear above the left during an error or tip are translated. Fixed minor bugs and edits. Also translated several objects that are mapped, when created.