Translated, update readme, backup
Translated the names of the animals and the description to them and in the database information.
Information on plants and their indices during cultivation in captivity and in domestic growth.
Some information about buildings and facilities has been translated, but not all.
Some lines in translation have been corrected and bug fixes corrected.